move options tickers and expirations to helpers
