+++ openbb_sdk/providers/intrinio/openbb_intrinio/models/options_chains.py
@@ -18,6 +18,8 @@ from openbb_provider.standard_models.options_chains import (
 from openbb_provider.utils.helpers import make_request
 from pydantic import Field, validator
 
+from openbb_intrinio.utils.helpers import Options
+
 TICKER_EXCEPTIONS = [
     "SPX",
     "XSP",
@@ -31,10 +33,6 @@ TICKER_EXCEPTIONS = [
     "XAU",
     "OEX",
 ]
-
-intrinio_session = requests_cache.CachedSession(
-    "OpenBB_Intrinio", expire_after=timedelta(days=5), use_cache_dir=True
-)
 
 
 class IntrinioOptionsChainsQueryParams(OptionsChainsQueryParams):
@@ -109,63 +107,6 @@ class IntrinioOptionsChainsFetcher(
     """Transform the query, extract and transform the data from the Intrinio endpoints"""
 
     @staticmethod
-    def get_options_tickers(api_key: str) -> List[str]:
-        """Returns all tickers that have existing options contracts.
-
-        Parameters
-        ----------
-        api_key: str
-            Intrinio API key.
-
-        Returns
-        -------
-        List[str]
-            List of tickers
-        """
-
-        url = f"https://api-v2.intrinio.com/options/tickers?api_key={api_key}"
-
-        r = intrinio_session.get(url, timeout=10)
-
-        if r.status_code != 200:
-            raise RuntimeError("HTTP Error -> Status Code: " + str(r.status_code))
-
-        return r.json()["tickers"]
-
-    @staticmethod
-    def get_options_expirations(
-        symbol: str,
-        api_key: str,
-        after: Optional[Union[datetime, str]] = datetime.now().strftime("%Y-%m-%d"),
-        before: Optional[str] = "",
-    ) -> List:
-        """Returns a list of all current and upcoming option contract expiration dates for a particular symbol.
-
-        Parameters
-        ----------
-        symbol: str
-            The options symbol, corresponding to the underlying security.
-        api_key: str
-            Intrinio API key.
-        after: str
-            Return option contract expiration dates after this date. Format: YYYY-MM-DD
-        before: str
-            Return option contract expiration dates before this date. Format: YYYY-MM-DD
-        """
-
-        url = (
-            f"https://api-v2.intrinio.com/options/expirations/{symbol}/"
-            f"eod?before={before}&after={after}&api_key={api_key}"
-        )
-        r = make_request(url)
-
-        if r.status_code != 200:
-            raise RuntimeError("HTTP Error -> Status Code: " + str(r.status_code))
-
-        expirations = pd.DatetimeIndex(sorted(list(r.json()["expirations"])))
-        return list(filter(lambda x: x > pd.to_datetime(after), expirations))
-
-    @staticmethod
     def generate_url(symbol, expiration, date, api_key):
         return f"https://api-v2.intrinio.com/options/chain/{symbol}/{expiration}/eod?date={date}&api_key={api_key}"
 
@@ -188,7 +129,7 @@ class IntrinioOptionsChainsFetcher(
         symbol: str, date: Optional[Union[str, dateType]] = "", api_key: str = ""
     ):
         symbol = symbol.upper()
-        SYMBOLS = IntrinioOptionsChainsFetcher.get_options_tickers(api_key)
+        SYMBOLS = Options.get_options_tickers(api_key)
         if symbol not in SYMBOLS:
             raise RuntimeError(f"{symbol}", "is not supported by Intrinio.")
 
@@ -199,7 +140,7 @@ class IntrinioOptionsChainsFetcher(
         if symbol in TICKER_EXCEPTIONS:
             symbol = "$" + symbol
 
-        expirations = IntrinioOptionsChainsFetcher.get_options_expirations(
+        expirations = Options.get_options_expirations(
             symbol=symbol, api_key=api_key, after=date
         )
         urls = []
@@ -225,8 +166,8 @@ class IntrinioOptionsChainsFetcher(
             results = IntrinioOptionsChainsFetcher.download_json_files(urls)
 
         df = pd.DataFrame()
-
-        for i in range(0, len(results)):
+        results_range = range(0, len(results))
+        for i in results_range:
             chains = results[i]
 
             contract_symbol: List[str] = []
